Split active-exam bug regression integration tests
Complete DEBT-370 Phase 1 by splitting the active-exam regression
families into projection/count/status coverage and latest-attempt /
fallback coverage, then removing the oversized grab-bag file.

All 24 bug-regression integration tests are preserved by name, and each
new sibling remains below the 1,200 LOC Phase 1 target.
